requirements.txt
discord.py[voice]
opuslib